Gather flowr runtime functions into a directory
