+++ flowr/src/runtime/args/get.rs
@@ -3,7 +3,7 @@ use std::env;
 use flow_impl::{DONT_RUN_AGAIN, Implementation, RunAgain};
 use serde_json::Value;
 
-use super::super::FLOW_ARGS_NAME;
+use ::FLOW_ARGS_NAME;
 
 #[derive(Debug)]
 /// `Implementation` struct for the `get` function
@@ -29,8 +29,8 @@ mod test {
 
     use flow_impl::{DONT_RUN_AGAIN, Implementation};
 
+    use super::FLOW_ARGS_NAME;
     use super::Get;
-    use super::super::super::FLOW_ARGS_NAME;
 
     #[test]
     fn test_arg_passing() {
